deep linking working. refactoring send page
