+++ lib/ui/sideDrawer.dart
@@ -385,8 +385,8 @@ class _SideDrawer extends State<SideDrawer>
 
   resetFn() async {
     services<AppRouter>().replaceAll([InitialPage()]);
-    services<DBManager>().deleteDatabase();
     services<WalletsService>().resetService();
+    services<DBManager>().deleteDatabase();
   }
 
   /// creates a button that display a dialog to choose an item and peek at the active item
